NETOBSERV-227 fix filtering by namespace or owner
https://issues.redhat.com/browse/NETOBSERV-227

- filters on "SrcK8S_OwnerName" or "DstK8S_OwnerName" shouldn't trigger
  common filters processing
- "Namespace" was renamed "K8S_Namespace"
- Use more generic solution to retrieve CommonLabelFilter out of
  configured labels
- side change: move CSV related functions to a different package to
  improve readability
